gpu.c: add_to_from_device: further clarify nature of domain and prefix arguments
The domain and the prefix arguments are already clarified not to have
been affected by the contraction that may have been applied during
the creation of the kernel, but it should not be affected by
any other contraction either.  Adjust the caller to take into account
any possible such other contraction.
Note that the schedule trees from which the statement instances
are extracted currently do not contain any contractions, but this may
change when statement instances are combined during scheduling.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>