gpu: handle mapping index expressions of arrays of structs to device memory
PPCG currently does not map any arrays containing structures to
shared or private memory, but in an upcoming commit, some of
these arrays may get mapped to private memory.
Change the application of the tiling to the index expressions
in transform_index to take into account that the index expression
may refer to some subfield of an outer array, while the tiling
always applies to an outer array.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>