gpu: allow arrays of structs to be (forcibly) mapped to private memory
In an upcoming commit, some particular form of arrays of structures
will be allowed to be mapped to private memory.  In particular,
the order constraints on these arrays will be removed from
the coincidence constraints, which may require them to be mapped
to private memory.  If this is the case, then actually allow
them to be mapped to private memory by creating the appropriate
private memory tiles.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>