gpu_array_info: store copy of declared extent
The declared extent is needed to derive AST expression for the declared size.
This information was obtained directly from the corresponding pet_array,
but this assumes a one-to-one correspondence between the arrays in a gpu_prog
and the arrays in the corresponding pet_scop.  While there currently
is such a correspondence, this will no longer be maintained
in an upcoming commit.
Store a copy of the declared extent in gpu_array_info instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>