ppcg_print_declaration_with_size: do not take pet_array as argument
ppcg_print_declaration_with_size is called from within
gpu_print_local_declarations and in order to get access
to the appropriate pet_array, it needs to assume that there
is a one-to-one correspondence between the arrays in a gpu_prog
and the arrays in the corresponding pet_scop.  While there currently
is such a correspondence, this will no longer be maintained
in an upcoming commit.
Since ppcg_print_declaration_with_size only needs the pet_array
for the element type of the array and since this information
is also available in the gpu_array_info, pass the element type
directly to ppcg_print_declaration_with_size.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>