gpu_group_references: do not construct shared tile on force_private
When an array is marked force_private, compute_group_bounds_core
always constructs a private tile.  It may however also construct
a shared tile, even if the array is marked force_private.
This did not cause problems before because the private tile
would be selected instead of the shared tile.
However, since b926320 (gpu_array_ref_group_type: take tile
with minimal depth if both types exist, Mon Jun 29 11:17:53 2015 +0200),
a shared tile may end up getting selected if it has a smaller depth.
In case of a force_private array, the computed depth of a shared tile
may not be meaningful, so do not construct the shared tile
in the first place to avoid this problem.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>