add intra-statement consecutivity constraints for consecutive arrays
For each statement that accesses arrays marked consecutive,
add intra-statement consecutivity constraints that aim to achieve
consecutivity on all accesses to the array from that statement.
If any access cannot be made consecutive (e.g., because
it has non-affine index expressions or an innermost index
expression that is not linearly independent of outer index
expressions), then ignore all accesses to the array from the statement.

If a statement accesses multiple arrays marked consecutive,
then construct intra-statement consecutivity constraints
that try and combine as many as possible.
This results in a list of intra-statement consecutivity constraints
where each entry covers the same number of arrays or more arrays
than the next entries.  The scheduler then tries each alternative
in turn.

Consecutivity constraints are currently only added for
the C target as it is mainly meant for demonstration purposes.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>