add inter-statement consecutivity constraints for consecutive arrays
For each pair of statement instances from distinct statements
that access consecutive elements (either both reading
the elements or both writing them) of an array that is marked
consecutive, add an inter-statement consecutivity constraint
that tries to cause the instances to be executed consecutively.
Do this for each index expression level from innermost to outermost,
but only if the two accesses do not get killed in between and
if the corresponding intra-statement consecutivity constraints
can be constructed.

Consecutivity constraints are currently only added for
the C target as it is mainly meant for demonstration purposes.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>