dump schedule constraints that are used as input to the isl scheduler
The main purpose of dumping schedule constraints is to see
what input is sent to the scheduler.
However, if grouping is applied, then the dumped schedule constraints
would be those from before the grouping.
Reorganize ppcg_compute_schedule such that dumping of schedule
constraints is performed in a single location, right before
the call to isl_schedule_constraints_compute_schedule.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>