gpu_group.c: set_stride: return indication of whether stride was found
This allows the caller to use this information directly instead
of having to figure it out indirectly from bound->stride.
This will make is easier to always store a stride in the next commit,
even if it is trivial.
Note that the caller could also check if this stride is trivial
or not, but the callee already performs this check, so it might
as well return this information.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>