gpu_group.c: can_tile: separate stride detection from bound computation
In particular, instead of looking at each individual dimension
in the access and determining a stride and a bound for each
separately, first determine (and remove) all strides and
only then compute bounds.
This will make it easier to promote the entire bound computation
functionality to isl.

The code also gets simplified a bit and it could be further simplified
by changing the way the shifts and the strides are stored.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>