take into account contraction when generating OpenMP support
Commit ppcg-0.05-103-gc7d7a176 (optionally group chains of statements,
Thu Mar 17 17:06:40 2016 +0100) introduced the possibility
of grouping several statements into a single statement for scheduling,
but failed to update the detection of parallel loops
in the OpenMP support accordingly.

In particular, when statements are grouped together,
the schedule tree refers to the groups rather than the individual
statements, while the dependence relations still refer
to the original statements.
The practical result is that all dependences get ignored and
every loop is considered parallel when grouping takes place.

Reformulate the partial schedule used in the detection
of parallel loops in terms of the original statements
such that the dependences do get taken into account.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>